src/conversiones-requirements.txt
# Conversiones Microservice Dependencies
# Siguiendo patrón tutorial-8-sagas

# FastAPI y servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Base de datos
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography>=41.0.0,<46.0.0

# Messaging
pulsar-client==3.3.0

# Utilidades
python-multipart==0.0.6
python-json-logger==2.0.7

# Testing (opcional)
pytest==7.4.3
httpx==0.25.2